Fix references and organize
