+++ src/pages/stats/abbreviations.astro
@@ -1,6 +1,6 @@
 ---
 import Default from '@layouts/Default.astro';
-import { getAbbreviations } from '@scripts/data';
+import { getAbbreviations } from '@scripts/data/abbreviations';
 
 const abbreviations = await getAbbreviations();
 ---
