+++ src/pages/stats/players/[id].astro
@@ -1,6 +1,7 @@
 ---
 import Default from '@layouts/Default.astro';
-import { getPlayers, getGames } from '@scripts/data';
+import { getGames } from '@scripts/data/games';
+import { getPlayers } from '@scripts/data/players';
 import { getFormattedDatetime } from '@scripts/utils';
 
 // Create page for each player
