+++ src/pages/stats/players/index.astro
@@ -1,6 +1,6 @@
 ---
 import Default from '@layouts/Default.astro';
-import { getPlayers } from '@scripts/data';
+import { getPlayers } from '@scripts/data/players';
 import { getFormattedDatetime } from '@scripts/utils';
 
 const players = await getPlayers();
