+++ src/pages/stats/teams/[club_id]/[season].astro
@@ -1,6 +1,7 @@
 ---
 import Default from '@layouts/Default.astro';
-import { getGames, getTeamId, getTeams } from '@scripts/data';
+import { getGames } from '@scripts/data/games';
+import { getTeamId, getTeams } from '@scripts/data/teams';
 import { getFormattedDatetime } from '@scripts/utils';
 
 // Create page for each team
