+++ src/pages/stats/teams/[club_id]/index.astro
@@ -1,6 +1,6 @@
 ---
 import Default from '@layouts/Default.astro';
-import { getClubs } from '@scripts/data';
+import { getClubs } from '@scripts/data/teams';
 
 // Create page for each team
 export async function getStaticPaths() {
