+++ src/pages/stats/teams/index.astro
@@ -1,6 +1,6 @@
 ---
 import Default from '@layouts/Default.astro';
-import { getTeams } from '@scripts/data';
+import { getTeams } from '@scripts/data/teams';
 
 const teams = await getTeams();
